# speedup with pmin and do.call ref #20