``` r
m <- matrix(sample(1:20,20000,replace=TRUE),ncol=5)
bench::mark(
    apply = apply(m, 1, min),
    pmin = do.call(pmin, as.data.frame(m))
)
#> # A tibble: 2 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 apply        2.68ms   2.82ms      347.     156KB     30.3
#> 2 pmin       129.33µs 133.66µs     7027.     253KB     21.1
```

<sup>Created on 2023-11-16 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>

this is probably as good as it gets